Consolidate BPM logic to custom hook - moves logic from App.tsx to useBPM.ts
